% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love_cats.R
\name{love_cats}
\alias{love_cats}
\title{Do you love cats?}
\usage{
love_cats(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
love_cats()
}
\keyword{cats}
